restrict caching to appropriate files

+++ literaryHalifax/www/js/cacheLayer.js
@@ -14,8 +14,6 @@ angular.module('literaryHalifax')
             // cache for the results of GET requests
             // hash url->object
             itemCache = {},
-            // cache which stores fi;enames of images, audio, etc.
-            fileCache = {},
             // whether or not to automatically cache GET requests
             // TODO this should be a persistent setting
             cacheIncoming = false,
@@ -23,8 +21,6 @@ angular.module('literaryHalifax')
             rootDir = '',
             // filename of the JSON dump of the item cache
             itemCacheFile = '',
-            // filename of the JSON dump of the file cache
-            fileCacheFile = '',
             // remote locations of files and items
             api = localization.resources.serverAddress + "api/",
             files = localization.resources.serverAddress + "files/",
@@ -40,18 +36,26 @@ angular.module('literaryHalifax')
             };
 
 
+        // The following two functions perform a mapping from urls on the server to valid local filenames
+        // a one-to-one mapping failed for some reason, so this only works for files that follow the format
+        // serverAddress/files/word/word.extension
+    
+        // if caching audio for directions is broken, this is probably why.
+    
         // convert a url into a filename
         function hash(url) {
             var urlParts = url.split("/"),
-                tmp = urlParts.pop();
-            return (urlParts.pop() + '-' + tmp);
+                tmp = urlParts.pop(),
+                result = urlParts.pop() + '-' + tmp;
+            return result;
         }
 
         // convert a filename into a url
         function unhash(filename) {
             var parts = filename.split('-'),
-                tmp = parts.pop();
-            return files + parts.pop() + '/' + tmp;
+                tmp = parts.pop(),
+                result = files + parts.pop() + '/' + tmp;
+            return result;
         }
     
         function applyParams(records, params) {
@@ -116,14 +120,46 @@ angular.module('literaryHalifax')
                 // we've found the files. Check for cached versions
                 lodash.forOwn(raw.file_urls, function (value, key) {
                     if (value) {
-                        raw.file_urls[key] = fileCache[value] || raw.file_urls[key];
+                        var newUrl = hash(value);
+                        // if rootDir is undefined, there is no possibility of a cached file
+                        if (rootDir) {
+                            promises.push(
+                                // check if the file is cached
+                                $cordovaFile.checkFile(rootDir, newUrl)
+                                    .then(function () {
+                                        // if it is cached, replace the url
+                                        $log.info("found file " + rootDir + '/' + newUrl);
+                                        raw.file_urls[key] = rootDir + '/' + newUrl;
+                                        return rootDir + '/' + newUrl;
+                                    }).catch(function () {
+                                        $log.info("could not find file " + rootDir + '/' + newUrl);
+                                        // otherwise, use the real url
+                                        // TODO: if we implement airplane mode, replace it with
+                                        // a placeholder image instead
+                                        return value;
+                                    })
+                            );
+                        }
                     }
                 });
-                promises = [$q.when()];
                 
             } else if(raw.directions_url) {
-                raw.directions_url = fileCache[raw.directions_url] || raw.directions_url;
-                promises = [$q.when()];
+                var newUrl = hash(raw.directions_url);
+                        // if rootDir is undefined, there is no possibility of a cached file
+                        if (rootDir) {
+                            promises.push(
+                                // check if the file is cached
+                                $cordovaFile.checkFile(rootDir, newUrl)
+                                    .then(function () {
+                                        // if it is cached, replace the url
+                                        raw.directions_url = rootDir + '/' + newUrl;
+                                        return rootDir + '/' + newUrl;
+                                    }).catch(function () {
+                                        // otherwise, use the real url
+                                        return raw.directions_url;
+                                    })
+                            );
+                        }
             }
             return $q.all(promises).then(function () {
                 return raw;
@@ -171,7 +207,7 @@ angular.module('literaryHalifax')
         // if the url is falsy, return true (null is cached because we don't
         // need to make a request to use it)
         function isCachedUrl(url) {
-            return !url || url.startsWith(rootDir) || fileCache[url];
+            return !url || url.startsWith(rootDir);
         }
     
         // Dump the item cache to a file
@@ -195,17 +231,6 @@ angular.module('literaryHalifax')
                 });
         }
 
-        function recoverFileCache() {
-            return $cordovaFile.readAsText(rootDir, fileCacheFile)
-                .then(function (result) {
-                    fileCache = angular.fromJson(result);
-                });
-        }
-    
-        function saveFileCache() {
-            return $cordovaFile.writeFile(rootDir, fileCacheFile, angular.toJson(fileCache), true);
-        }
-
         // delete the item cache
         function destroyItemCache() {
             return $cordovaFile.removeFile(rootDir, itemCacheFile)
@@ -213,13 +238,6 @@ angular.module('literaryHalifax')
                     itemCache = {};
                     status.cacheEnabled = false;
                 });
-        }
-
-        // delete the file cache
-        function destroyFileCache() {
-            return $cordovaFile.removeFile(rootDir, fileCacheFile).then(function () {
-                fileCache = {};
-            });
         }
     
         // download and cache one item type
@@ -306,23 +324,20 @@ angular.module('literaryHalifax')
             return $cordovaFileTransfer
                 .download(url, filename)
                 .then(function (success) {
-                    // try to write this change to the file, but if we fail there's not a lot we can do
-                    saveFileCache().catch(function (error) {
-                        $log.error("could not save file cache: " + error);
-                    });
-                    fileCache[url] = filename;
+                $log.info("saved file " + filename);
                     return filename;
                 });
         };
 
         // uncache the given path
         layer.clearUrl = function (path) {
-            
-            var originalUrl = lodash.findKey(fileCache, function (fileName) {
-                return fileName == path;
-            }),
-                filename = path.split('/').pop();
-            
+            if (!path) {
+                $log.warn("Tried to clear a null path");
+                return $q.when(path);
+            }
+
+            var filename = path.split("/").pop(),
+                url = unhash(filename);
             return $cordovaFile.checkFile(rootDir, filename)
                 .then(function () {
                     return $cordovaFile.removeFile(rootDir, filename);
@@ -330,15 +345,9 @@ angular.module('literaryHalifax')
                     //the file already does not exist
                     return $q.when();
                 }).then(function (success) {
-                    fileCache[originalUrl] = undefined;
-                    // try to write this change to the file, but if we fail there's not a lot we can do
-                    saveFileCache().catch(function (error) {
-                        $log.error("could not save file cache: " + error);
-                    });
-                
                     // for convenience, resolve with the url that
                     // the cached file came from
-                    return originalUrl;
+                    return url;
                 });
         };
 
@@ -356,25 +365,27 @@ angular.module('literaryHalifax')
 
         // delete every cached file, then delete the item cache
         layer.destroyCache = function () {
-            var filesIndex = itemCache[layer.getRequest('files').url],
-                promises = [];
+            var filesIndex = itemCache[layer.getRequest('files').url];
             status.working = true;
-                
             // decorate the file cache so that remote urls are replaced with their cached versions.
             // Normally modifying the cache like this is bad, but
             // it's going to be deleted anyway
-            
-            lodash.each(fileCache, function (cachedUrl) {
-                promises.push(layer.clearUrl(cachedUrl));
-            })
-            return $q.all(promises).then(
-                $q.all([
-                    destroyItemCache(),
-                    destroyFileCache()
-                ]
-                )).finally(function () {
-                status.working = false;
-            });
+            return decorate(filesIndex)
+                .then(function () {
+                    var promises = [];
+                    lodash.each(filesIndex, function (file) {
+                        
+                        lodash.forOwn(file.file_urls, function (value, key) {
+                            if (value && isCachedUrl(value)) {
+                                promises.push(layer.clearUrl(value));
+                            }
+                        });
+                    });
+
+                    return $q.all(promises).then(destroyItemCache);
+                }).finally(function () {
+                    status.working = false;
+                });
         };
 
         // This is a bit out of place. We retrieve files from the
@@ -433,7 +444,6 @@ angular.module('literaryHalifax')
         $ionicPlatform.ready(function () {
             rootDir = "cordova.file.dataDirectory";
             itemCacheFile = 'itemCache';
-            fileCacheFile = 'fileCache'
 
             if (typeof cordova !== 'undefined') {
                 rootDir = cordova.file.dataDirectory;
@@ -453,7 +463,7 @@ angular.module('literaryHalifax')
             
             $cordovaFile.checkFile(rootDir, itemCacheFile)
                 .then(function (success) {
-                    return recoverItemCache().then(expandIndices).then(recoverFileCache);
+                    return recoverItemCache().then(expandIndices);
                 }, function (error) {
                     // no cache, that's fine
                     return $q.when();
